feat: Add AffirmationsPractice screen and Meditate screen
The code changes include adding a new screen called AffirmationsPractice and a new screen called Meditate. The AffirmationsPractice screen displays affirmations and allows the user to navigate through them. The Meditate screen displays a meditation image. These changes were made to enhance the functionality and user experience of the application.